[chore] 빌드오류 수정
